app/requirements.txt
fastapi[standard]>=0.115.12
motor
pydantic[email]
pydantic-settings
pyjwt
pwdlib[argon2]
python-multipart
pytest>=8.0.0
pytest-cov>=5.0.0
httpx>=0.27.0
flake8>=7.0.0
black>=24.0.0
bandit>=1.7.0
mkdocs>=1.6.0
mkdocs-material>=9.5.0
mkdocs-mermaid2-plugin>=1.1.0
mangum
boto3